+++ dependency/lpp/periodic_session/assistance_data.cpp
@@ -66,14 +66,15 @@ bool AssistanceDataHandler::request_assistance_data() {
     VSCOPE_FUNCTION();
 
     messages::RequestAssistanceData message_description{};
-    message_description.cell               = mData.cell;
-    message_description.periodic_session   = handle();
-    message_description.gps                = mData.gnss.gps;
-    message_description.glonass            = mData.gnss.glonass;
-    message_description.galileo            = mData.gnss.galileo;
-    message_description.bds                = mData.gnss.beidou;
-    message_description.delivery_amount    = mData.config.delivery_amount;
-    message_description.rtk_antenna_height = mData.config.antenna_height;
+    message_description.cell                        = mData.cell;
+    message_description.periodic_session            = handle();
+    message_description.gps                         = mData.gnss.gps;
+    message_description.glonass                     = mData.gnss.glonass;
+    message_description.galileo                     = mData.gnss.galileo;
+    message_description.bds                         = mData.gnss.beidou;
+    message_description.delivery_amount             = mData.config.delivery_amount;
+    message_description.rtk_antenna_height          = mData.config.antenna_height;
+    message_description.disable_update_capabilities = mData.config.disable_update_capabilities;
 
     if (mData.type == PeriodicRequestAssistanceData::Type::OSR) {
         message_description.rtk_observations           = 1;
@@ -108,14 +109,15 @@ bool AssistanceDataHandler::update_assistance_data(supl::Cell cell) {
 
     // TODO(ewasjon): remove the duplication code
     messages::RequestAssistanceData message_description{};
-    message_description.cell               = cell;
-    message_description.periodic_session   = handle();
-    message_description.gps                = mData.gnss.gps;
-    message_description.glonass            = mData.gnss.glonass;
-    message_description.galileo            = mData.gnss.galileo;
-    message_description.bds                = mData.gnss.beidou;
-    message_description.delivery_amount    = mData.config.delivery_amount;
-    message_description.rtk_antenna_height = mData.config.antenna_height;
+    message_description.cell                        = cell;
+    message_description.periodic_session            = handle();
+    message_description.gps                         = mData.gnss.gps;
+    message_description.glonass                     = mData.gnss.glonass;
+    message_description.galileo                     = mData.gnss.galileo;
+    message_description.bds                         = mData.gnss.beidou;
+    message_description.delivery_amount             = mData.config.delivery_amount;
+    message_description.rtk_antenna_height          = mData.config.antenna_height;
+    message_description.disable_update_capabilities = mData.config.disable_update_capabilities;
 
     if (mData.type == PeriodicRequestAssistanceData::Type::OSR) {
         message_description.rtk_observations           = 1;